# api/swagger/v1/todo-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Todo service",
    "version": "1.0",
    "contact": {
      "name": "go-grpc-http-rest-microservice project",
      "url": "https://github.com/hamitb/go-grpc-http-rest-microservice",
      "email": "hamitburakemre@gmail.com"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/todo/all": {
      "get": {
        "summary": "Read all todos",
        "operationId": "ReadAll",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ReadAllResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist",
            "schema": {
              "format": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "api",
            "description": "API versioning.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "TodoService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "v1ReadAllResponse": {
      "type": "object",
      "properties": {
        "api": {
          "type": "string",
          "title": "API Versioning"
        },
        "todos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Todo"
          },
          "title": "List of all todos"
        }
      }
    },
...
    "v1Todo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64",
          "title": "Unique integer identifier"
        },
        "title": {
          "type": "string",
          "title": "Title of the task"
        },
        "description": {
          "type": "string",
          "title": "Detailed description of the todo"
        },
        "reminder": {
          "type": "string",
          "format": "date-time",
          "title": "Date and time to remind the todo"
        }
      },
      "title": "Todo structure"
    },
...
  }
}